Move setting input selection to the end to helpers.
